refacator: 세마포어 락 관리 및 해제 AOP 적용과 작업의 수 관리 별도의 클래스로 분리
