+++ src/main/java/com/map/gaja/client/application/geocode/Geocoder.java
@@ -2,7 +2,6 @@ package com.map.gaja.client.application.geocode;
 
 import com.map.gaja.client.infrastructure.file.parser.dto.ParsedClientDto;
 import com.map.gaja.client.infrastructure.api.KakaoGeoApi;
-import com.map.gaja.client.application.geocode.exception.LockAcquisitionFailedException;
 import com.map.gaja.client.application.geocode.exception.TooManyRequestException;
 import com.map.gaja.client.application.geocode.exception.NotExcelUploadException;
 import lombok.RequiredArgsConstructor;
@@ -14,10 +13,7 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 import java.util.List;
-import java.util.concurrent.Semaphore;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import static com.map.gaja.client.constant.LocationResolverConstant.*;
 
@@ -27,29 +23,16 @@ import static com.map.gaja.client.constant.LocationResolverConstant.*;
 public class Geocoder {
     private final KakaoGeoApi kakaoGeoApi;
     private final ResponseParser responseParser;
-    private final Semaphore semaphore = new Semaphore(1);
-    private final AtomicInteger totalTaskCount = new AtomicInteger(0);
-
 
     /**
      * 비동기 통신으로 도로명 주소를 위경도로 변환
      */
     public Mono<Void> convertToCoordinatesAsync(List<ParsedClientDto> addresses) {
-        int taskCount = addresses.size();
-
-        increaseTaskCount(taskCount);
-
-        acquireLock(taskCount);
-
         return Flux.fromIterable(addresses)
                 .filter(this::hasAddress)
                 .delayElements(DELAY_ELEMENTS_MILLIS)
                 .timeout(TIMEOUT_SECONDS)
                 .flatMap(this::convert)
-                .doOnTerminate(() -> {
-                    decreaseTaskCount(taskCount);
-                    semaphore.release();
-                })
                 .doOnError(this::handleError)
                 .then();
 
@@ -60,15 +43,6 @@ public class Geocoder {
                 .flatMap(responseParser::parse)
                 .doOnNext(parsedClientDto::setLocation)
                 .then();
-    }
-
-    /**
-     * 통신을 해야 되는 작업이 많을 경우 사용자는 오래 기다려야 되고 자원을 점유하는 시간도 길어지므로 빠른 시간안에 서비스를 이용할 수 있는지 확인한다.
-     */
-    public void checkServiceAvailability() {
-        if (isLongWait()) {
-            throw new TooManyRequestException();
-        }
     }
 
     private boolean hasAddress(ParsedClientDto data) {
@@ -83,29 +57,4 @@ public class Geocoder {
         }
     }
 
-    private void acquireLock(int taskCount) {
-        try {
-            if (!semaphore.tryAcquire(LOCK_TIMEOUT, TimeUnit.SECONDS)) {
-                throw new LockAcquisitionFailedException();
-            }
-        } catch (Exception e) {
-            decreaseTaskCount(taskCount);
-            throw new NotExcelUploadException(e);
-        }
-    }
-
-    private void increaseTaskCount(int taskCount) {
-        totalTaskCount.addAndGet(taskCount);
-    }
-
-    private void decreaseTaskCount(int taskCount) {
-        totalTaskCount.addAndGet(-taskCount);
-    }
-
-    private boolean isLongWait() {
-        if (totalTaskCount.get() >= LIMIT_PROCESS_COUNT) {
-            return true;
-        }
-        return false;
-    }
 }
